requirements.txt
Django==6.0.4
djangorestframework>=3.14,<3.16
webcolors==25.10.0
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
